De-alias types in TypeTagIdentifier
De-alias types in `TypeTagIdentifier` so that a binding
for a type alias can be found by an identifier for that
alias's value.

Due to restrictions in 2.11, the `TypeTagIdentifier.apply` cannot
be overridden; as a result, this changes the class to not to be a
case class. As part of this change which may not be binary
compatible, the version is bumped to 0.6.0-SNAPSHOT.